+++ assets/javascripts/application.js
@@ -301,7 +301,7 @@
   };
 
   maxdown = {
-    version: '0.3.7 (2. July 2015)',
+    version: '0.3.8 (7. July 2015)',
     cm: '',
     autosave_interval_id: null,
     autosave_interval: 5000,